requirements.txt
Flask
psutil
wmi
speedtest-cli
pywin32
